Don't modify SONAME for libnvjpeg.so.9.0
This is a hacky workaround specific to NVJPEG.

Prior to this change, when importing nvidia.dali, we would see this error:

ImportError: /usr/local/lib/python2.7/dist-packages/nvidia/dali/.libs/libnvjpeg-63a2f672.so.9.0: version `libnvjpeg.so.9.0' not found (required by /usr/local/lib/python2.7/dist-packages/nvidia/dali/.libs/libdali-5fa9fbd1.so)

The reason is that CUDA libraries like libnvjpeg.so.9.0 set their .gnu.version
info in the ELF... to their own name.  which means that if we go messing around
with the SONAME, then anything that ever linked against that library will be
broken.  For the rest of CUDA we don't have to deal with this because we rely
on the CUDA libs to be installed as system libs.