Switch over completely to FindProtobuf.cmake from CMake 3.9.6 (#41)
Previously we had imported only a fraction of the FindProtobuf.cmake from CMake 3.9.6, but future improvements will be much cleaner if we just use the whole thing from upstream. Only two tiny changes to include() statement paths are necessary to get the later FindProtobuf to work when dropped in on top of CMake 3.5.