Fix for low accuracy in Tensorflow training
Indexed_file_loader.h performs a sharded read of the index files supplied alongside the tfrecords for training.
Thus, ReadSample() is called from multiple threads with different shard_id_. When all indexes have been read, we
proceed to reset current_index_ to start again at the beginning of the shard.

This commit fixes a defect in the detection that the shard has arrived to the end, by comparing the value of
current_index to the size of all the indexes. The fix consists in comparing against the end of the shard instead
of the end of the indices.

Before the fix, the code would work correctly for 1 GPU when there is only 1 shard, but with >1 GPUs, it would
result in a loss in accuracy due to---most likely---redundant reads.

Signed-off-by: Pablo Ribalta Lorenzo <pribalta@nvidia.com>