Change Reader to allow for custom-type handling
Add LoadTarget to Loader and Parser.
It is intended to allow type-safe sample loading,
as user can provide custom wrapper, where
metadata can be stored safely.

This removes UB from file_loader - int label for image
was placed in a u8 buffer with image data,
and was accessed by reinterpreting a unaligned pointer.

Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>