Add fine grained control over output buffers in the pipeline (#212)
Add finer control over buffers returned from the pipeline

With current asynchronous executor design, we have n-buffers, while 1
is being consumed, other can be calculated. When n is equal to 2
and we release old one and start consuming new one. If released
one cannot be filled during one network forward and backward pass
everything waits until it is filled. By this path, it is possible to
immediately release the buffer as soon as data is copied to some external
tensor. Also, this change adds the ability to set pipeline length in the
executor - the default value is 2.

Signed-off-by: Janusz Lisiecki <jlisiecki@nvidia.com>