Add Clang to Dockerfile.deps and pass CC and CXX as arguments
This allows user to specify build-args to Dockerfile
for CC=clang and CXX=clang++ and build with clang.

To workaround issue with is_trivially_copyable
not present in libstdc++ from CI, check for specific value
of __GLIBCXX__ was added. Otherwise it is hard to detect it
when compiling with clang.

Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>