Fix race condition in AsyncPipelinedExecutor destructor
When AsyncPipelinedExecutor is destroyed WorkerThread destructor is called as well. Inside, it waits for the running job to be completed and then thread executing it is shut down as well. The problem arises because the job that is executed
inside WorkerThread uses conditionals and mutexes that are destroyed during AsyncPipelinedExecutor destruction and before WorkerThread finishes.
So WorkerThread may end up hanging inside work() it needs to complete before
is able to shutdown, but this work it trying to use no longer existing
conditional or mutex waiting for it infinitely.

Signed-off-by: Janusz Lisiecki <jlisiecki@nvidia.com>