Deliver exactly 1 epoch from DALIGenericIterator
If ```stop_at_epoch``` is set to True, DALIGenericIterator will deliver exactly ```size``` data entries in total by returning a fractional batch of what was loaded onto the GPU and processed in the last iteration. Otherwise, it will behave as in the past, returning an integer multiple of ```self._num_gpus * self.batch_size``` entries, which will always exceed ```size``` if ```size``` is not divisible by ```self._num_gpus```.

Signed-off-by: Paul Bisso <41169757+paulbisso@users.noreply.github.com>
Signed-off-by: Ubuntu <ubuntu@ip-172-31-47-183.us-east-2.compute.internal>