Handle NULL pointer in ctypes_void_ptr
Since a94190ca4bcfcdb94802c3e74b4259fe51e5f58f, a CUDA stream can be
passed to `copy_to_external`.

When passing the default CUDA stream to this function, for instance
with `ctypes.c_void_p(0)`, this parameter is converted to the `None`
Python singleton, and thus is not a valid integer value that can be
passed to `PyLong_AsVoidPtr`. This was causing `copy_to_external` to
return with the exception "TypeError: an integer is required" being
set.

Signed-off-by: Felix Abecassis <fabecassis@nvidia.com>