Fix a race condtion in async pipeline executor (#1103)
* Check for pending shutdown in AsyncPipelineExecutor threads.
* Notify executor threads upon shutdown.

Background:
When DALI pipeline is shut down it is possible that only GPU part of scheduled work is picked up by the worker thread while preceding mixed stage is not executed and worker thread end before producing any result. DALI is shutting down, and GPU worker thread is not able to finish waiting on the conditional variable for mixed stage results. This PR adds a break of this wait in the async pipeline executor destructor

Signed-off-by: Janusz Lisiecki <jlisiecki@nvidia.com>