Upgrade VideoCodecSDK to 9.0.20
Description: Upgrade VideoCodecSDK to the latest 9.0.20 release.
This patch maintains the runtime loading behaviour with libnvcuvid.so
but removes the dynlink_* headers that have been deprecated by Video
codec library and are not reccommended anymore by the codec team.
Future upgrades will be as easy as dropping in the new headers (nvcuvid.h
and cuviddec.h) and declaring function pointers that will be loaded
at runtime.

Signed-off-by: Abhishek Sansanwal <asansanwal@nvidia.com>