Improve heuristic for variable frame rate detection
Description: Previous heuristic would fail for videos with
a fractional fps, such as 29.97(30000/1001). It would classify
such videos as vfr even if they might be cfr.
This change will compare the frame rate reported by the container
to an estimate based off on the first video frame.
While this check is not exhaustive but in my testing it was accurate.

Signed-off-by: Abhishek Sansanwal <asansanwal@nvidia.com>