Improve heuristic for variable frame rate detection (#1242)
Description: Previous heuristic would fail for videos with
a fractional fps, such as 29.97(30000/1001). It would classify
such videos as vfr even if they might be cfr.
This change will compare the frame rate reported by the container
to an estimate based off on the first video frame.
While this check is not exhaustive, but in my testing it was accurate.

Why we need this PR?
Refactoring to improve what
Improves vfr detection heuristic.
What happened in this PR?
Explain solution of the problem, new feature added.
Compare the frame rate reported by the container
to an estimate based off on the first video frame
What was changed, added, removed?
Replaced old check with new check for vfr.
Was this PR tested? How?
Tested with constant frame rate video with 29.97 fps
Were docs and examples updated, if necessary?
No

Signed-off-by: Abhishek Sansanwal <asansanwal@nvidia.com>
Signed-off-by: Jianjun Liu <00liujj@163.com>