Add tf dataset (#1299)
Why we need this PR?
It adds new feature needed to improve DALI TensorFlow support and prepare it for the upcomming TF 2.0
What happened in this PR?
Explain solution of the problem, new feature added.
To support future versions of TensorFlow there is a need to use tf.data.Dataset API rather then the older options of feeding the data to the TensorFlow graph.
This PR adds initial support for this API in DALI.

What was changed, added, removed?
Added DALIDatasetOp class and its Python wrappers. Modified how DALI TF plugin is build to include those.

What is most important part that reviewers should focus on?
How DALI API is used in the context of TensorFlow DatasetOp.

Was this PR tested? How?
Added python tests for DALIDataset.

Were docs and examples updated, if necessary?
Yes

JIRA TASK: [DALI-380]

Signed-off-by: Albert Wolant <awolant@nvidia.com>
Signed-off-by: Jianjun Liu <00liujj@163.com>