Add support for constants in expressions (backend)
Add support for constant inputs to expressions.
ConstantStorage keeps the constants
and allows for easy access to them through type-erased
pointers.
Use common factory for GPU and CPU.
CPU and GPU use the same Extended Tiles with
type-erased input and output pointers.
This allows to get rid of one layer of templates
and reduce the compilation time significantly.

Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>