Fix scalar batch handling in arithmetic ops
Adjust test for checking if the Tensor should
be considered a scalar.
Allow Scalar batch to be broadcasted in the operation
properly.

Return shape is always a batch - fixed for Constant/Scalar
inputs.

Add C++ unit tests for Scalar Batch.
Test if Pipeline behaves properly when switching between
scalar and non-scalar inputs.

Python test was extended to cover Scalar inputs,
size of some of the tests was reduced.

Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>