Fix race condition between consecutive invocations of stage
Some of the operators, like MakeContiguous
use internal CPU buffers as inputs to CUDA kernels.

Consecutive invocation of the stage may allow
the CPU part of the operator to overwrite the
contents of such buffer before the async CUDA
call even starts, resulting in mismatch of the data
between iterations.

This commits makes the Mixed and GPU stages
(the ones that have a stream and use CUDA kernels)
wait with the start of the next iteration for
all the CPU and GPU work from previous one to finish.

Additionaly the EventPool was reworked to create
Events without ever accepting the max limit,
as reusing Events may break some assumptions in DALI.

Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>