Move pipeline ownership from Dataset to Iterator
According to `MakeIterator` docs:

"This method may be called multiple times on the same instance,
and the resulting iterators will have distinct state. Each
iterator will traverse all elements in this dataset from the
start."

The Dataset stores the information of how to create
instance of Pipeline, creates it when necessary
and hands it down to Iterator.
Every Iterator will have a separate Pipeline
that it will destroy when the destructor is invoked.

Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>