Fix Transpose bugs - degenerate dims and non-uniform GPU
cuTT which is used as backend for DALI Transpose
cannot handle degenerate dimensions equal 1,
so they are removed from shape description
together with the corresponding entries from permutation.

As the permutation describes transformation:
Destination dimension i <- source dimension perm[i]

If we remove dimension `i` from the shape,
we need to remove entry of a value `i` from perm,
and reduce the netries that are bigger.

It was done the other way round, removing the
entry at position `i` in perm.

Additionaly, cuTT cannot handle a permutation
with only one dimension - we now fall back
to a cudaMemcpyAsync.

In case of shape = {1, 1, ..., 1}, the preparation
step reduced it to empty shape -> this is now
special cased to always reduce to shape equal {1}
and perm = {0}, which fixes a bug in CPU Transpose
that didn't check for an empty case.

Non-uniform batch case for GPU was also fixed,
as it was not accesing the elements of the batch,
but only the first element.

GTest test were extended with additional checks
for the shape & perm reduction (for degenerate dims = 1),
and for non-unfiorm batch shapes.

Additional docstrings were added to Permutation CPU Impl.

Docstring was extended.

Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>