Fix semantics of the masks_meta field
As per the documentation of the COCOReader Ops, the masks_meta
output is defined as a list of tuples (mask_idx, start_idx, end_idx)
where start_idx indicates the index of the first coords in masks_coords.

Now it's not explicit what the meaning of end_idx is, but one could
reasonably expect it to be the index of the last coords in masks_coords.

It's not though, in the case where the annotation is not a polygon but a
list of polygons. This is actually reflected in the Reader's unit test :
for example the last annotation spans 2 polygons, and the masks_meta for
them is
3, 28, 38,
3, 38, 34,
where the end_idx of the second annotation is smaller than its
start_idx.

I took the liberty to align the semantics with the documentation, since
it seems to me that it's much easier to interpret that way.

Signed-off-by: Nicolas Lopez <nicolas.lopez@mines-paris.org>