CUDA VM memory resource (#3114)
`cuda_vm_resource` - a memory resource that combines an general purpose coalescing free-tree pool with a paging-like mechanism that maps regions of physical storage to a large pre-allocated virtual memory arena. When there's free mapped memory in the free tree, it is allocated. Otherwise, a virtual memory region is obtained and physical blocks are mapped to it. The physical blocks may be reassigned from a free region or allocated anew, if there are no free blocks to spare.

Tests:
- basic test
- random allocation
- targeted tests for virtual address region merging
- targeted tests for block remapping

Signed-off-by: Michał Zientkiewicz <mzient@gmail.com>