Make graph-mode test ultra fast
It seems that with every test in graph mode
the default graph is growing and takes longer
and longer to execute. We just clear the default
graph before running the tests and now the
whole test_dali_tf_dataset_graph executes orders
of magnitude faster, almost matching the eager
mode tests.

Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>