Fix infinite loop for convolution with extent equal 1
GPU kernel's border handling went into infinite loop,
as the border reflect 101 is not suited for inputs
with extent zero. As the CPU version treats it as
border replicate, we do the same for GPU.

As it boils down to `sum(window) * input_element`
we compact the window to 1 element and break the
loop early.

Test to cover edge cases added.

Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>