Separate type setting from data access in Buffer
Remove `set_type<T>` from `mutable_data<T>`.
This is the only accessor, that can also change
the underlying type of the data and/or cause
an implicit allocation.

Change most of the calls to set the type
with the `Resize(shape, type)` call
and in template contexts with `set_type<T>`.

Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>