Fix data race when copying outputs in TF plugin
DALI TF Plugin was copying the outputs on
TF's stream and was not synchronizing with
the copy.
In some cases we can overtake the copy
with DALI iteration and overwrite the buffers
before they are actually copied.

Add a flag to synchronize with the stream after
last copy is scheduled so we synchronize only
once.

Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>