Add non-owning SampleView
The purpose is to have Tensor wrapper that is similar
to already existing DALI vocabulary types representing
Tensor:
* it is non-owning view to the data
* it allows to access the pointer, the shape and the type id
* can be used as argument to view the same way as Tensor
but as opposed to Tensor doesn't allow to adjust the allocation.

The target of this type is to use it as the return
type of the TensorBatch::operator[] - using static type
is not possible due to the Tensor[List/Vector] not being
statically typed. This should allow for proper encapsulation
of the batch object - SampleView allows access to the data
but does not allow to adjust the allocation or changing
the metadata through the view - as oposed to the currently
used Tensor.

Add conversion from SampleView to statically typed TensorView
using view<T, ndim> template function.

Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>