Introduce runtime typed TensorView
The new specialization uses `DynamicType` tag
as the type parameter to TensorView.

The purpose is to have Tensor wrapper that is similar
to already existing DALI vocabulary types, and additionally
allows to keep the type as a runtime information.

The target of this type is to use it as the return
type of the TensorBatch::operator[] - using static type
is not possible due to the Tensor[List/Vector] not being
statically typed. This should allow for proper encapsulation
of the batch object - TensorView allow access to the data
but do not allow to adjust the allocation or changing
the metadata through the view - as oposed to the currently
used Tensor.

Additionaly extend the utility of the `view` function
template to allow for any conversions between
statically and dynamically typed TensorViews.

Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>