Fix TF DALIDataset tests that changed layout between iterations (#3836)
DALI assumes that the input will have the dimensionality
inferred from the layout passed to external source operator.

TF DALIDataset allows to specify input and override
the previously specified layout - which could change
the dimensionality. As changing the dimensionality
is not allowed this leads to error in tests.

TODO: Adjust the API by either disallowing
to override the layout or allow it only when initial
layout is empty. Introduce better error checking
that would detect that situation.

Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>