Add .style.yapf file
As we already have a .clang-format for C++, this PR
introduces corresponding configuration for yapf,
which allows to format Python.

It is similar to clang-format as it allows to reformat
existing code that already passes the lint (as opposed
to other formatters like autopep8).

It is also quite configurable and allows to be applied
to parts of the file, contrary to black, allowing
to be used selectively.

This configuration conforms to the currently used style
within the project.

Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>