Fix and reenable optimized Cast kernel
The bug in binsearch Cast kernel manifested
on empty samples - there was mismatch between
sample that was assigned to a given block
and the block that was assumed by the kernel
due to BlockDescriptos not being generated
for samples with 0 volume.

Now, the empty samples are skipped at the
end of the batch and when assigning the
lowest blockIdx to a given sample.

The binsearch looks for the rightmost
sample when searching the descriptors.

Test was adjusted to not generate empty samples
and another was added to randomly make
selected ranges of samples empty.

The speedup compared to the old implementation
is between 10% and 25%, averaging 16%, depending
on the size of the input.

Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>