Expose device_id in ShareData and SetSample APIs
The device_id is necessary metadat when accepting
external memory via pointers.

The device_id can differ from the one that is
present in the order. device_id represents the
location of memory, the order tracks the necessary
synchronization - the provided memory can in theory
depend on a work being done on another device.

TODO:
* Add this mention to docstrings
* Add a test that checks Python Tensor/TensorList
constructors in their current state (conversion from CuPy?)

Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>