Adjust Doxygen configuration
Replace the usage of `///` with `/** */` which is used
across the project.

Unify the usage of grouping directives.
The header should use either just `/** @{ */`
or additionaly @name or @defgroup.

  @name can be used for member functions, groups them,
  and keeps them inlined on class page.

  @defgroup can be used with free function or friend
  definitions, or other top level entities. It generates
  a dedicated page for the group.

The foother should use `/** @} */`

Set DISTRIBUTE_GROUP_DOC = YES - now the first docstring
in a group is replicated in the other members that didn't
provide a docstring.

Set EXTRACT_ALL = YES - now, all entities are automatically
documented - otherwise, for example, free functions were
ignored in file without @file documentation.
This enables most free-functions (for example C API)
and not just Classes and their members to be documented.

Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>