Add AsTensor and related APIs to Tensor Vector
Remove the atomic count of views, as it didn't
add any value, and propagating samples with
the counter might have weird effects.

The cacheing is removed - but the API still
returns a view. As the DALI types are currently
not modifiable in place from Python level,
the lifetime should be mostly the same.

Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>