Add AsTensor and related APIs to Tensor Vector (#4124)
Add to the TensorVector the set of APIs already
present in TensorList:
* IsContiguousTensor
* IsDenseTensor
* AsReshapedTensor
* AsTensor

The APIs allow to access the memory of the contiguous
batch through the API of the Tensor. They are exposed
through the Python APIs for TensorList and the conversion
is valid as long as the allocation of the source object
is not changed.
Caching the returned Tensors is not replicated to simplify
the API.

The atomic counter of current internal "view" Tensors 
(the elements of tensors_) is removed - it use was insignificant
and the propagation of samples with such counter might
have strange side effects.  

There will be a difference with how the Tensors will be returned 
to Python: instead of keeping the instance of the created
Tensor as a member of the batch class and returning
a pointer (like TensorList does), we return the Tensor
object with a shared_ptr that co-owns the allocation
- which handles the issue of allocation lifetime when
the batch is gced but the Tensor is not.

Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>