Refactor Tensor Vector (#4149)
Refactor TensorVector with intention of using it as the
only batch object through DALI.

Main changes:
1. Batch is the source of truth. 
This breaks away from the design, where depending on the case
(contiguous vs noncontiguous) our metadata were kept
in the member TensorList or in the first Tensor respectively.
This is enabled thanks to better encapsulation where
the internal allocation cannot be changed from the outside.

2. When modifying the batch, care was taken to update the properties
(now stored in batch) in the same order as they are defined.
There are some exceptions. 

3. Now setting and copying samples can be verified against
all the properties of the batch - they must match. 

4. There are now three options when deciding if the batch
should use contiguous allocation.
We can either pin the contiguous or noncontiguous modes
or allow buffer to change depending on the needs.
In the last case, it can be specifically requested during
Resize to change to one of the modes (but without pinning it). 

5. When Resizing the batch, and a new allocation is needed,
we try to use one big allocation if allowed by the selected mode.
It yields smaller memory fragmentation 

6. When we convert the buffer to non-contiguous and it had
a contiguous backing allocation, we mark the tensors_
that were pointing to that allocation as not sharing, but still aliasing.
This is so we can call Resize() and get a new allocation without
the error that we try to reallocate something that shares data.

5. Contents of `tensor_list_as_tensor_vector_test.cc`
are copied from `tensor_list_test.cc` and adjusted to handle
both contiguous and noncontiguous cases.
It is necessary to verify the TensorVector conformance
to the TensorList requirements but the code is almost identical
to the already existing one. 

The tensor_vector.h was documented and reordered
into groups of methods.

Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>