Use std::exchange in move operator for Tensors
This adjusts Tensor and SampleView types,
so they both have the empty {0} shape after
move.
Previously it left the Tensor with a scalar shape
that could indicate existing allocation after
move.

As the use after move shouldn't be practised,
this change should not impact anything.

TensorViews and TensorShapes are left mostly
unaffected as the static_dim doesn't compose
nicely with forcing {0} shape or clearing
the dimensionality. For a TensorView, it is
valid to have a shape and a nullptr,
for the shapes they can have dimensionality
and no samples.

Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>