Remove TensorList and use only TensorVector
Introduces an type alias that replaces
TensorList with TensorVector.

Next step will be to rename all the
batch object to uniform name and update
the naming of UnsafeCopy and UnsafeSet APIs.

Limit the PassThrough to only 1-1 mappings,
as it is currently used.
It will be extended for the upcoming split
and merge operations.

Introduce a pass in executor that detects
when the input to MakeContiguous is always
contiguous to allow for PassThrough.

PassThrough for the MakeContgiuous needs
to be computed to allow for correct sizing
of prefetch buffer queues - otherwise
we would waste the memory when we need
to do the copy by prefetching the inputs
needlessly, or we would ask executor to allocate
the output of MakeContiguous and than share
data into it causing an immidaite deallocation.

MakeContiguous is now inserted before every output,
so we make sure that all outputs are contgiuous.
Previously it was guaranteed that GPU outputs
were contgiuous due to the use of TensorList.

Possible followup is to enable outputs from
CPU stage again.

ArgumentWorkspace uses TensorVector only now.

All redundant overloads taking TensorList were
removed, leaving only the more generic TensorVector
ones.

Some tweaks to debug utilities.

Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>