Fix the indexing for newer PyTorch
In new PyTorch when using one tensor to do selection from another
tensor, the indicied are required to be on the same device or on the
cpu. Previously, we indexed a CPU tensor with GPU indexes to get a GPU
tensor as a result. This change transfers the indexed tensor to GPU
first, so both tensors are on the same device and still return a GPU
tensor.

Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>