Rename TensorVector to TensorList
TensorList was replaced by a refactored TensorVector that supports both
the TensorList APIs as well as non-contiguous mode in which in can be
constructed from individual samples. The TensorList was kept as an
alias to TensorVector to limit the code changes.
As the TensorList name is more prelevant across DALI codebase, TensorVector
is renamed into TensorList and the alias is removed, leaving us with just
one batch type.

The tests are merged into one file, taking the ones from original
TensorVector and the ones that were ported to it from TensorList tests.
Old TensorList tests are removed as they were less generic.

Changes are fully automated, both the search&replace as well as code
formatting.

Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>