Support tiled TIFFs (#4201)
Until now, only striped TIFFs were supported. In this PR I add support for tiled TIFFs.

Additional information:
The main disadvantage of striped TIFFs is that efficient ROI decoding is effectively impossible - you always need to read a whole strip. This is the reason why Tiled TIFFs are common in applications where the data is bigger, for example in spatial information or microscopy. That's why we want to support them.

Before, this decoder was reading TIFF line-by-line. Now I extended it to support tiles and, in order to avoid code duplication, I simply pretend that lines are 1 x imagewidth tiles.

We do not support exotic bitdepths (i.e. other than 8, 16 and 32), because I was unable to find software capable of correctly writing or reading those, so it was impossible to test reliably.

Signed-off-by: Szymon Karpiński <skarpinski@nvidia.com>