Support applying orientation in Convert (#4219)
Images may be stored in an orientation other than the actual one and contain a tag (see EXIF orientation) indicating that the image should be transformed (for example rotated) after decoding.

In this PR I extend Convert function with an ability to rectify an image according to the orientation information.

Additional information:

The implemented solution is not the fastest one, but is definitely the simplest: I adjust the strides, sizes and pointer of the output tensor in such a way, that the data is stored there in appropriate order. A function responsible for that is `ApplyOrientation`.

Not all decoders use the high-level Convert, some (such as TIFF) use lower-level version. In this
case `ApplyOrientation` can be used directly in order to modify the strides, sizes and pointer.

Signed-off-by: Szymon Karpiński <skarpinski@nvidia.com>