Improve error handling in LibTiffDecoder (#4210)
While testing LibTiffDecoder, I've identified two major problems with error handling there:

Corrupted (or exotic, unsupported) files will cause an exception, which would kill the pipeline. In this PR I add a bulk catch to gracefully handle those and allow for a fallback instead of killing the pipeline
Not enough checks were made at the beginning and as a result TIFFs that we couldn't decode correctly (planar or with non-standard photometric interpretation) were not rejected. In this PR I add checks for those

Signed-off-by: Szymon Karpiński <skarpinski@nvidia.com>