Fix libtiff error handling (#4274)
Two bug fixes:

- Throwing in C code 
When using libtiff, I was passing a set of helper functions (read, seek etc.) for it to read the input. When invalid offset was passed to seek, it was throwing an exception. However, the libtiff is a C library, so throwing there was UB. I wrap the seek in a try-catch to fix that.

- Leaking memory on failed open
When TIFFClientOpen failed, the allocated DecoderHelper was not deleted, causing a leak.

Signed-off-by: Szymon Karpiński <skarpinski@nvidia.com>