Fix inconsistent tensor recreation in TensorList
Use the same shape for the internal tensor that gets back into the scope
when the size of the TensorList is changed. empty_shape can be only used
for non-zero dimensional samples, as 0d samples can't have 0 volume.

The code uses an explicit TensorShape<0> converted to one of dynamic
dimension, as the brace default constructed TensorShape<> had size 1 for
some reason.

This clears up potential asserts that could hit in debug builds:
specifically calling empty_shape for dim = 0.

Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>