Aggregating ImageDecoder (#4224)
## Aggregating decoder

The ImageDecoder implements the ImageDecoderInstance interface, just as specific decoders do. It also implements ImageParser, because it can also parse arbitrary images.

ImageDecoder aggreates several decoders, filtered by a predicate. These decoders are then selected based on compatibility with image format and priority,

Operation:
For each decoder, there's a worker thread which picks up a sub-batch of work and calls ScheduleDecode on the associated decoder, then it waits results.
If a decoder has a fallback, then the worker waits for partial results. If there are failures, the failing samples are sent to the fallback worker. The successes are propagated to the top-most result promise.
If a decoder doesn't have a fallback, the worker waits for all results and propagates them to the top-most results promise, when ready.

Signed-off-by: Michal Zientkiewicz michalz@nvidia.com
Co-authored-by: Joaquin Anton janton@nvidia.com
Co-authored-by: Kamil Tokarski <kamiltokarski04@gmail.com>