Add Split and Merge and adjust executor to support them
Split and merge only accept the boolean predicate now.

Executor detects empty batches, resets the output and doesn't run the
operator as they were not designed to handle empty inputs.

The executor does best effort to pin everything that goes into merge, if
one of the inputs was pinned so we have consistent batch merge
operation.

For split and merge the check for equal batch size in input and output
is disabled.

Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>