Extend and relax TensorList sample APIs (#4358)
Add ResizeSample that allows to resize individual samples. Automatically
converts the batch into non-contiguous mode, same as SetSample.

SetSample no longer requires the order between the batch and new sample
to match - instead it waits for the incoming order in the order of
current batch, so the access in the batch order will be correct.

Fix a problem that prohibited copying from empty batch that had no
samples nor type set. Now it resets the current batch to similar state.

Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>