Add internal Split and Merge operators (#4359)
Add samplewise passthrough information to graph - it informs the graph, 
that the operator may pass through data on any-to-any basis (specifying 
the mapping would not be practical if we want to enable more than 
two-value selection - similar to a `switch` rather than an `if`).
It also indicates that the data won't be guaranteed to be passed through
without changes, so the executor can't consider it as passing through
contiguous data.

Add Split and Merge operators. 
Both accept only a boolean predicate that describes the split, but can
be easily extended to support multiple values in a switch-like fashion.

Split divides samples into two outputs based on the predicate. All the
metadata regarding the input (dimensionality, type, pinnedness, etc)
is passed unchanged to both outputs.

Merge does the opposite to Split. The problems arise around inputs that
can have different order or pinnedness - the first can happen around
stage boundary and with pass-through operators, the latter might be
sometimes out of our control when external source is involved.

The order problem is solved/worked around by implicit synchronization
that was introduced in SetSample and is now transparent for this
operator.

For pinned memory (that is a problem only for CPU op), the executor
tries to detect if we pinned any of the inputs to the Merge node, 
and if so, tries to pin all the inputs (including following through 
the pass-through). 
The Merge op will treat the output as pinned if executor marked the
outputs as pinned, or if any of the inputs turns out to be pinned.
It also needs to utilize this information to set the correct device_id.

Downgrading the status of memory from pinned to regular was considered
at some point, but based on the recent synchronization problems with
Xavier, it was decided on copy to a buffer of correct pinned status,
so not to get surprised by suddenly fully async copies from supposedly
non-pinned memory.

In optimal case, when everything is consistently pinned and the input
orders are consistent the merge operation should not introduce any
additional synchronization or copies.

Executor had to be extended and now detects empty batches, resets
the output and doesn't run the operator as they were not designed
to handle empty inputs.
External source prevents empty batches from entering already existing
pipelines, so the old behaviour of not supporting them is preserved.

For Split and Merge the check for equal batch size in input and output
is disabled.

Operators are introduced into `_conditional` module (with a leading
underscore) and without exposing their documentation (at least for now).

Fix the accesses to schema when inserting the operator into fn API, 
so we know that the doc is hidden, an it is introduced via the `.hidden`
module and must be reimported into module above. 

Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>