Fix ExecutorSyncTest that run the SimpleExecutor twice
The graph was executed twice, after that we looked at the result. For
SimpleExecutor this means that the data can be already overwritten, we
got lucky with timing and we managed to check the first result before it
got wiped by the second one.

Adjust the test, so for simple executor it checks the data immediatelly
after it is produced.

Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>