Replace Status::OK() with Status() for TF plugin (#4442)
TensorFlow introduced a Tensorflow Support Library, where the Status was
moved. 
https://github.com/tensorflow/tensorflow/commit/6a7f86b78612ed909058b4656b37b761abc2559f
removed the `Status::OK()` static function, advocating the usage of 
`OkStatus()` or the default constructor `Status()` which is supposed
to be backward compatible. As we are targeting TF 1.15 as well, where
the `OkStatus()` is some TF-lite wrapper, we go with the `Status()` as
a replacement.

Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>