Execute whole Keras code in the expected device scope
In newer TF version, when the model.fit is executed without explicit
scope (which gives it default placement) we may encounter transfers
between the device on which we defined the model and inputs to the
device which executes the model - which is not supported. Adjust those
tests to both be defined and execute on the desired device - so all the
tensors/variables placements are uniform.

Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>